fix babylon fp check
