+++ test/replay/finality_test.go
@@ -8,6 +8,7 @@ import (
 	appparams "github.com/babylonlabs-io/babylon/v3/app/params"
 	"github.com/babylonlabs-io/babylon/v3/testutil/datagen"
 	ftypes "github.com/babylonlabs-io/babylon/v3/x/finality/types"
+	ibctmtypes "github.com/cosmos/ibc-go/v10/modules/light-clients/07-tendermint"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -315,4 +316,34 @@ func TestMissingSignInfoNewlyActiveFpSet(t *testing.T) {
 
 	// vote and finalize a few more blocks
 	scn.FinalityFinalizeBlocksAllVotes(blkHeightLastVoted+1, 2)
+}
+
+func TestOnlyBabylonFpCanCommitRandomness(t *testing.T) {
+	t.Parallel()
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
+	driverTempDir := t.TempDir()
+	replayerTempDir := t.TempDir()
+	driver := NewBabylonAppDriver(r, t, driverTempDir, replayerTempDir)
+
+	consumerID1 := "consumer1"
+
+	// 1. Set up mock IBC clients for each consumer before registering consumers
+	ctx := driver.App.BaseApp.NewContext(false)
+	driver.App.IBCKeeper.ClientKeeper.SetClientState(ctx, consumerID1, &ibctmtypes.ClientState{})
+	driver.GenerateNewBlock()
+
+	// 2. Register consumers
+	consumer1 := driver.RegisterConsumer(consumerID1)
+	require.NotNil(t, consumer1)
+
+	// Create a Babylon FP (registered without consumer ID)
+	consumerFp := driver.CreateNFinalityProviderAccounts(1)[0]
+	consumerFp.RegisterFinalityProvider(consumerID1)
+	driver.GenerateNewBlock()
+
+	consumerFp.CommitRandomness()
+
+	txResults := driver.GenerateNewBlockAssertExecutionFailure()
+	require.Len(t, txResults, 1)
+	require.NotEqual(t, txResults[0].Code, uint32(0))
 }
